Creates Delete route to break MealFood relation for meals and foods
