+++ routes/api/v1/meals.js
@@ -19,6 +19,7 @@ router.post('/:mealId/foods/:foodId', function(req, res){
   mealFoodsController.create(req, res);
 });
 
+// Destroy - mealFoods
 router.delete("/:mealId/foods/:foodId", function(req, res){
   mealFoodsController.destroy(req, res)
 });
